requirements-docker.txt
loguru
omegaconf>=2.3.0
typer>=0.12.3
transformer-engine @git+https://github.com/NVIDIA/TransformerEngine.git@stable
dacite
humanize
wandb
ninja
--extra-index-url https://download.pytorch.org/whl/cu124 torchvision torchaudio torchao
torchdata @git+https://github.com/pytorch/data.git@v0.8.0
transformers
datasets
tiktoken
lm_eval @git+https://github.com/EleutherAI/lm-evaluation-harness.git@v0.4.3
fla @git+https://github.com/sustcsonglin/flash-linear-attention